VR Network Synchronization
The VR Rig is now synchronized over the network, it has been designed in such a way that a separate rig can exist for use in singleplayer menus, which will automatically be replaced with a separate rig connected to the network.
There are utilities attached to the VR rigs so that the same rig can be used in singleplayer, multiplayer local, and multiplayer controlled remotely environments; with the proper components being disabled on enabled in accordance with the current state.

The game currently automatically connects to the network at startup, this is the easiest method to debug it with.